--- keyboards/ergodox_ez/keymaps/bocaj/config.h
@@ -1,21 +1,0 @@
-#ifndef USERSPACE_CONFIG_H
-  #define USERSPACE_CONFIG_H
-
-  #ifndef QMK_KEYS_PER_SCAN
-    #define QMK_KEYS_PER_SCAN 4
-  #endif // QMK KEYS PER SCAN
-
-  // this makes it possible to do rolling combos (zx) with keys that
-  // convert to other keys on hold (z becomes ctrl when you hold it,
-  // and when this option isn't enabled, z rapidly followed by x
-  // actually sends Ctrl-x. That's bad.)
-  #define IGNORE_MOD_TAP_INTERRUPT
-  #define PERMISSIVE_HOLD
-  #undef PREVENT_STUCK_MODIFIERS
-
-  // Disable action_get_macro and fn_actions, since we don't use these
-  // and it saves on space in the firmware.
-  #define NO_ACTION_MACRO
-  #define NO_ACTION_FUNCTION
-
-#endif // !USERSPACE_CONFIG_H
