--- keyboards/ergodox_ez/keymaps/bocaj/readme.md
@@ -1,48 +1,0 @@
-# Ergodox EZ Layout by JacobJerrell
-
-My personal Ergodox EZ layout.
-
-## Summary
-
-Contains two base layers because I carry my keyboard between different machines and occassionally use the machines without this keyboard. A side-effect to using it in this manner (with a non-QWERTY layout) is that software/hardware mappings conflict so this allows a quick (2-buttons) to get the keyboard to a compatible layer
-
-## Layers
-
-| Layer | Purpose |
-| ----- | ------- |
-|   0   | [Hardware Workman](#hardware-workman) - For OS mapped to QWERTY |
-|   1   | [Software Workman](#software-workman) - For OS mapped to Workman |
-|   2   | [Lower](#lower) |
-|   3   | [Raise](#raise) |
-|   4   | [Adjust](#adjust) |
-|   5   | [Diablo III](#diablo-iii) |
-
-### Hardware Workman
-
-For easy plug and play into QWERTY machines
-![Hardware Workman](https://imgur.com/HHMIOEf.png)
-
-### Software Workman
-
-To switch to when I was previously using the laptop keyboard with software mapped to Workman. Or so other people can take the Ergodox for a spin.
-![Software Workman](https://imgur.com/WnSci7o.png)
-
-### Lower
-
-Reveals F1-12 keys and brings numkeys down.
-![Lowered](https://imgur.com/mGfjwcn.png)
-
-### Raise
-
-Reveals F1-F12 keys and gives shifted numkeys.
-![Raised](https://imgur.com/NC96Jus.png)
-
-### Adjust
-
-Numpad, Wipe EEPROM, OSX Volume Control, OSX Lockscreen, and Default Layer Switching.
-![Adjusted](https://imgur.com/aQtOxZU.png)
-
-### Diablo III
-
-Arranged to make playing on a laptop much more enjoyable. Credit for the D1-4 macros (as well as much of the ideas in this layout goes to Drashna)
-![Diablo III](https://imgur.com/5eQcDOQ.png)
