+++ layouts/community/ergodox/bocaj/config.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include QMK_KEYBOARD_CONFIG_H
+
+#undef PRODUCT
+#define PRODUCT   BocajEZ - Frankensteined ErgoDox EZ
