+++ layouts/community/ergodox/bocaj/rules.mk
@@ -3,3 +3,4 @@ COMMAND_ENABLE = no
 SWAP_HANDS_ENABLE = no
 TAP_DANCE_ENABLE = yes
 EXTRAKEY_ENABLE = no
+KEY_LOCK_ENABLE = yes
