+++ users/bocaj/rules.mk
@@ -1,0 +1,5 @@
+SRC += bocaj.c
+
+ifeq ($(strip $(TAP_DANCE_ENABLE)), yes)
+  SRC += tap_dances.c
+endif
